+++ src/Controller/UserController.php
@@ -11,6 +11,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 /**
  * @Route("/admin/user")
@@ -97,6 +98,9 @@ class UserController extends AbstractController
      */
     public function delete(Request $request, User $user): Response
     {
+        if($user->isSuperAdmin()) {
+            throw new AccessDeniedException('Cannot delete super admin.');
+        }
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($user);
